+++ CHANGES.rst
@@ -1,4 +1,9 @@
+Unreleased
+==========
+
+* implement ``PipFreeze.__bool__()`` (py3) and ``PipFreeze.__nonzero__()`` (py2)
+
 0.0.1
 =====
 
-Implement ``pipfreeze.PipFreeze``.
+* implement ``pipfreeze.PipFreeze``