+++ CHANGES.rst
@@ -1,5 +1,5 @@
-Unreleased
-==========
+0.0.2
+=====
 
 * implement ``PipFreeze.__bool__()`` (py3) and ``PipFreeze.__nonzero__()`` (py2)
 * implement ``PipFreeze.__iter__()``